requirements.txt
Django==4.0.4
djangorestframework==3.13.1
django-cors-headers==3.12.0
djoser==2.1.0

